// src/app/shared/types/api-output-types/administration-contact/contact-list-output-dto.interface.ts
export interface ContactListOutputDTO {
  contactName?: string;
  firstName?: string;
  lastName?: string;
  customerCode?: string;
  title?: string;
  phone1?: string;
  phone2?: string;
  fax?: string;
  email?: string;
  address?: string;
  town?: string;
  country?: string;
  primaryContact?: boolean;
  statusState?: string;
  statusUser?: string;
  statusDate?: Date;
  businessPartner?: string;
}